+++ types/index.d.ts
@@ -262,6 +262,12 @@ declare namespace RamdaAdjunct {
         isNaN(val: any): val is typeof NaN;
 
         /**
+         * Checks if value is a natural number.
+         * Natural numbers correspond to all non-negative integers and 0.
+         */
+        isNaturalNumber(val: any): boolean;
+
+        /**
          * Checks whether the passed value is complement of `NaN` and its type is not `Number`.
          */
         isNotNaN(val: any): boolean;
@@ -536,8 +542,8 @@ declare namespace RamdaAdjunct {
             (either: Catamorphism<V1 | V2>,
             ) => T1 | T2;
         cata<V1, V2, T1, T2>(leftFn: (leftValue: V1) => T1): {
-            (rightFn: (rightValue: V2) => T1, either: Catamorphism<V1 | V2>): T1 | T2;
-            (rightFn: (rightValue: V2) => T1): (either: Catamorphism<V1 | V2>) => T1 | T2
+            (rightFn: (rightValue: V2) => T2, either: Catamorphism<V1 | V2>): T1 | T2;
+            (rightFn: (rightValue: V2) => T2): (either: Catamorphism<V1 | V2>) => T1 | T2
         };
 
         /**
@@ -554,6 +560,26 @@ declare namespace RamdaAdjunct {
          */
         renameKeysWith(renameFn: (key: string) => string, obj: object): object;
         renameKeysWith(renameFn: (key: string) => string): (obj: object) => object;
+
+        /**
+         * Creates a new object with the own properties of the provided object, but the
+         * key `key` renamed according to logic of renaming function.
+         */
+        renameKeyWith(
+            renameFn: (key: string) => string,
+            key: string,
+            obj: object
+        ): object;
+        renameKeyWith(
+            renameFn: (key: string) => string,
+            key: string
+        ): (obj: object) => object;
+        renameKeyWith(
+            renameFn: (key: string) => string
+        ): {
+            (key: string, obj: object): object;
+            (key: string): (obj: object) => object;
+        };
 
         /**
          * Create a new object with the own properties of the second object merged with
@@ -1191,7 +1217,7 @@ declare namespace RamdaAdjunct {
          * Creates an array with all falsy values removed.
          * The values false, null, 0, "", undefined, and NaN are falsy.
          */
-        compact<T>(list: T[]): Array<NonNullable<T>>;
+        compact<T>(list: T[]): Array<Exclude<NonNullable<T>, false | '' | 0>>;
 
         /**
          * Returns a new list containing the contents of the given list, followed by the given
@@ -1402,6 +1428,18 @@ declare namespace RamdaAdjunct {
         invokeArgs(pathToMethod: string[], args: Array<string | number>, obj: object): any;
         invokeArgs(pathToMethod: string[], args: Array<string | number>): (obj: object) => any;
         invokeArgs(pathToMethod: string[]): (args: Array<string | number>, obj: object) => any;
+
+        /**
+         * Converts double-precision 64-bit binary format IEEE 754 to signed 32 bit integer number.
+         */
+        toInteger32(n: number): number;
+        toInt32(n: number): number; // alias
+
+        /**
+         * Converts double-precision 64-bit binary format IEEE 754 to unsigned 32 bit integer number.
+         */
+        toUinteger32(val: number): number;
+        toUint32(val: number): number; // alias
     }
 }